requirements.txt
pyside6
pyqt6-tools
pyqtgraph
numpy
pillow
tesserocr
opencv-python